requirements.txt
Flask
PyPDF2
python-docx
nltk
scikit-learn
matplotlib
sentence-transformers
sqlite3
pandas